fixing mask projection

+++ util/utilIO.cpp
@@ -77,7 +77,7 @@ cv::Mat ImgIO::getIntrMatrix(const vcg::Shot<float> &shot){
   intr_mat.at<float>(1,1) = shot.Intrinsics.FocalMm;
   intr_mat.at<float>(0,2) = shot.Intrinsics.CenterPx[0];
   intr_mat.at<float>(1,2) = shot.Intrinsics.CenterPx[1];
-  intr_mat.at<float>(2,2,0) = 1.0f;
+  intr_mat.at<float>(2,2,0) = static_cast<float>(1.0);
 
   std::cout<< intr_mat <<std::endl;
